+++ swagger_gen/restapi/operations/variant/find_variants_urlbuilder.go
@@ -16,7 +16,7 @@ import (
 
 // FindVariantsURL generates an URL for the find variants operation
 type FindVariantsURL struct {
-	FlagID int32
+	FlagID int64
 
 	_basePath string
 	// avoid unkeyed usage
@@ -44,7 +44,7 @@ func (o *FindVariantsURL) Build() (*url.URL, error) {
 
 	var _path = "/flags/{flagID}/variants"
 
-	flagID := swag.FormatInt32(o.FlagID)
+	flagID := swag.FormatInt64(o.FlagID)
 	if flagID != "" {
 		_path = strings.Replace(_path, "{flagID}", flagID, -1)
 	} else {
